Restructuring database configuration 5/x. Still work in process, but unit and integration tests all OK
