+++ rabbit-core/src/main/java/org/ohdsi/databases/configuration/ConfigurationValidator.java
@@ -1,0 +1,6 @@
+package org.ohdsi.databases.configuration;
+
+@FunctionalInterface
+public interface ConfigurationValidator {
+    ValidationFeedback validate(ConfigurationFields fields);
+}
