+++ rabbit-core/src/main/java/org/ohdsi/databases/configuration/DBConfigurationException.java
@@ -1,0 +1,7 @@
+package org.ohdsi.databases.configuration;
+
+public class DBConfigurationException extends RuntimeException {
+    public DBConfigurationException(String s) {
+        super(s);
+    }
+}
