+++ rabbit-core/src/main/java/org/ohdsi/databases/configuration/FieldValidator.java
@@ -1,0 +1,6 @@
+package org.ohdsi.databases.configuration;
+
+@FunctionalInterface
+public interface FieldValidator {
+    ValidationFeedback validate(ConfigurationField field);
+}
